refactor: Router 의존관계 변경
- Router 생성 후 export하고 App 에서 pageComponents들을 할당하도록 변경
