+++ .prettierrc
@@ -12,7 +12,6 @@
   "requirePragma": false,
   "semi": true,
   "singleQuote": true,
-  "tabWidth": 2,
   "trailingComma": "es5",
   "useTabs": false,
   "vueIndentScriptAndStyle": false
