+++ src/js/index.js
@@ -1,6 +1,7 @@
 import '../css/index.css';
 import { $ } from './utils/DOM';
 import App from './App';
+import Router from './Router';
 
 const initalState = {
   isLogin: false,
@@ -9,6 +10,7 @@ const initalState = {
 const app = new App({
   parentNode: $('#app'),
   state: initalState,
+  Router: Router,
 });
 
 app.addStaticEventListeners();
